moving auth.js to routes/
